api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
